+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## Features
 
+- Improve progress messages on console
 - Add alternative names for `NODE_MIRROR`: `NVM_NODEJS_ORG_MIRROR`,
-  `N_NODE_MIRROR` and `NODIST_NODE_MIRROR`.
+  `N_NODE_MIRROR` and `NODIST_NODE_MIRROR`
 
 # 2.0.0